chore: move get-note-pipe to api utils
